Implemented plan to update money-related vars from double to big decimal

+++ app/src/main/java/com/example/expensetracker/data/Account.kt
@@ -4,6 +4,7 @@ import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.Index
 import androidx.room.PrimaryKey
+import java.math.BigDecimal
 
 @Entity(tableName = "accounts", indices = [Index(value = ["name"], unique = true)])
 data class Account(
@@ -12,5 +13,5 @@ data class Account(
     @ColumnInfo(collate = ColumnInfo.NOCASE)
     val name: String,
     val currency: String,
-    val balance: Double
+    val balance: BigDecimal
 )
